Move crc32_fold_copy helper functions into test harness.
These are not needed for the zlib implementation as it defaults to
crc32_fold_copy_c which calls the crc32 for the current architecture
and then memcpy.
