+++ arch/arm/crc32_armv8.c
@@ -72,15 +72,4 @@ Z_INTERNAL Z_TARGET_CRC uint32_t crc32_armv8(uint32_t crc, const uint8_t *buf, s
     return c;
 }
 
-/* Note: Based on generic crc32_fold_* implementation with functable call replaced by direct call. */
-Z_INTERNAL Z_TARGET_CRC void crc32_fold_copy_armv8(crc32_fold *crc, uint8_t *dst, const uint8_t *src, size_t len) {
-    crc->value = crc32_armv8(crc->value, src, len);
-    memcpy(dst, src, len);
-}
-
-Z_INTERNAL Z_TARGET_CRC void crc32_fold_armv8(crc32_fold *crc, const uint8_t *src, size_t len, uint32_t init_crc) {
-    Z_UNUSED(init_crc);
-    crc->value = crc32_armv8(crc->value, src, len);
-}
-
 #endif
