+++ arch/loongarch/loongarch_functions.h
@@ -10,8 +10,6 @@
 
 #ifdef LOONGARCH_CRC
 uint32_t crc32_loongarch64(uint32_t crc, const uint8_t *buf, size_t len);
-void     crc32_fold_copy_loongarch64(crc32_fold *crc, uint8_t *dst, const uint8_t *src, size_t len);
-void     crc32_fold_loongarch64(crc32_fold *crc, const uint8_t *src, size_t len, uint32_t init_crc);
 #endif
 
 #ifdef LOONGARCH_LSX
@@ -45,10 +43,6 @@ void inflate_fast_lasx(PREFIX3(stream) *strm, uint32_t start);
 #  if defined(LOONGARCH_CRC)
 #    undef native_crc32
 #    define native_crc32 crc32_loongarch64
-#    undef native_crc32_fold
-#    define native_crc32_fold crc32_fold_loongarch64
-#    undef native_crc32_fold_copy
-#    define native_crc32_fold_copy crc32_fold_copy_loongarch64
 #  endif
 #  if defined(LOONGARCH_LSX) && defined(__loongarch_sx)
 #    undef native_adler32
