+++ functable.c
@@ -248,8 +248,6 @@ static int init_functable(void) {
 #ifdef ARM_CRC32
     if (cf.arm.has_crc32) {
         ft.crc32 = &crc32_armv8;
-        ft.crc32_fold = &crc32_fold_armv8;
-        ft.crc32_fold_copy = &crc32_fold_copy_armv8;
     }
 #endif
 
@@ -315,8 +313,6 @@ static int init_functable(void) {
 #ifdef LOONGARCH_CRC
     if (cf.loongarch.has_crc) {
         ft.crc32 = crc32_loongarch64;
-        ft.crc32_fold = &crc32_fold_loongarch64;
-        ft.crc32_fold_copy = &crc32_fold_copy_loongarch64;
     }
 #endif
 #ifdef LOONGARCH_LSX
